fixed date not sorted ascendingly and removed time from date in output excel
had to change date to datetime dtype first to enable sorting to work in collated_df, and then change back to str to remove time from date in output excel(excel automatically adds time to date when the date has a datetime dtype). This also solves the issue with merging files with different date format in raw data (yy and yyyy).